Fix E2E Tests (#31205)
* Removed Changlogger Autoload

* Changed Jetpack Changelogger Formatter Loading

Rather than using the Composer autoload, we should use the
changelogger's "filename" option to point directly at the file.
This keeps the file out of the autoloads, since it's possible that
plugins or packages may be symlinked and break the relative
path to the plugin.

Committed via a GitHub action: https://github.com/woocommerce/woocommerce/actions/runs/1464955706